Clear cart

+++ src/modules/cart/repository/cartRepository.js
@@ -1,8 +1,15 @@
 import Cart from "../../../databases/models/cart.js";
 
+
+
 const addProductToCart = async (data) => {
-  const newCart = new Cart(data);
+  const newCart = new Cart(data);  
   return await newCart.save();
+};
+
+
+const addManyProductsToCart = async (products) => {
+  return await Cart.insertMany(products); 
 };
 
 const updateCartQuantity = async (cartId, quantity) => {
@@ -22,16 +29,22 @@ const getCartByAttributes = async (key1, value1, key2, value2) => {
   return await Cart.findOne(query);
 };
 
-const deleteProductFromCart = async ( productId)=>{
-  return await Cart.findOneAndDelete ({ productId })
+const deleteProductFromCart = async ( userId, productId)=>{
+  return await Cart.findOneAndDelete ({ userId, productId })
 }
+
+const deleteAllProductFromCart = async (userId) => {
+  return await Cart.deleteMany({ userId }); 
+};
 
 
 export default {
   addProductToCart,
+  addManyProductsToCart,
   updateCartQuantity,
   getCartByAttribute,
   getCartByAttributes,
-  deleteProductFromCart
+  deleteProductFromCart,
+  deleteAllProductFromCart
   
 };
